pyproject.toml
[tool.poetry]
name = "stocks-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.2"
fastapi = { extras = ["standard"], version = "^0.115.2" }
pydantic-settings = "^2.6.0"
tenacity = "^9.0.0"
selenium = "^4.25.0"
beautifulsoup4 = "^4.12.3"
httpx = "^0.27.2"
fake-useragent = "^1.5.1"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.2" }
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
extend-select = [
    "I",   # isort
    "PL",  # Pylint
    "N",   # pep8-naming
    "ANN", # flake8-annotations
]

ignore = [
    "ANN002", # missing-type-args
    "ANN003", # missing-type-kwargs
    "ANN101", # missing-type-self
    "ANN102", # missing-type-cls
    "ANN401", # any-type
]
